+++ app/src/main/java/com/wakeforge/app/presentation/splash/SplashScreen.kt
@@ -269,7 +269,7 @@ private fun FloatingSplashParticles() {
 
     val time by infiniteTransition.animateFloat(
         initialValue = 0f,
-        targetValue = 2f * PI,
+        targetValue = (2 * PI).toFloat(),
         animationSpec = infiniteRepeatable(
             animation = tween(25000, easing = LinearEasing),
             repeatMode = RepeatMode.Restart,
